LPS-66930 Move roles related data handlers to the proper package
